Release version 1.7.2
