+++ tests/conftest.py
@@ -1,12 +1,11 @@
 """Pytest configuration and fixtures for Rxiv-Maker tests."""
 
+import gc
 import os
 import shutil
 import subprocess
 import tempfile
-import time
 from pathlib import Path
-from typing import Any, Dict, Optional
 
 import pytest
 
@@ -15,176 +14,49 @@ try:
     import sys
 
     sys.path.insert(0, str(Path(__file__).parent.parent / "nox_utils"))
-    from nox_utils import DiskSpaceMonitor, cleanup_manager
+    # from nox_utils import DiskSpaceMonitor, cleanup_manager
 
     CLEANUP_AVAILABLE = True
 except ImportError:
     CLEANUP_AVAILABLE = False
 
-# Global container registry for reuse detection
-_CONTAINER_REGISTRY: Dict[str, Dict[str, Any]] = {}
 
-
-def get_reusable_container(engine_name: str, image: str) -> Optional[str]:
-    """Check for existing reusable container."""
-    if not CLEANUP_AVAILABLE:
-        return None
-
-    try:
-        # Look for running containers with the same image
-        for engine in cleanup_manager.engines:
-            if engine.engine_type == engine_name:
-                containers = engine.get_containers(all_containers=False)  # Only running
-
-                for container in containers:
-                    # Check if container is suitable for reuse
-                    container_image = container.get("Image", "")
-                    container_id = container.get("ID", "")
-
-                    if image in container_image and container_id:
-                        # Check if container is healthy
-                        success, stdout, stderr = engine.run_command(["exec", container_id, "echo", "test"], timeout=5)
-                        if success:
-                            print(f"🔄 Reusing existing {engine_name} container: {container_id[:12]}")
-                            return container_id
-
-    except Exception as e:
-        print(f"⚠️  Container reuse check failed: {e}")
-
-    return None
-
-
-def register_container(container_id: str, engine_name: str, image: str, scope: str):
-    """Register container for cleanup tracking."""
-    _CONTAINER_REGISTRY[container_id] = {
-        "engine": engine_name,
-        "image": image,
-        "scope": scope,
-        "created_at": time.time(),
-        "workspace": str(Path.cwd()),
-    }
-
-
-def cleanup_container(container_id: str, engine_name: str, force: bool = False):
-    """Clean up a container with enhanced error handling."""
-    if not container_id:
-        return
-
-    try:
-        print(f"🧹 Cleaning up {engine_name} container: {container_id[:12]}")
-
-        # Try graceful stop first
-        subprocess.run([engine_name, "stop", container_id], capture_output=True, text=True, timeout=30)
-
-        # Force remove
-        subprocess.run([engine_name, "rm", "-f", container_id], capture_output=True, text=True, timeout=10)
-
-        # Remove from registry
-        if container_id in _CONTAINER_REGISTRY:
-            del _CONTAINER_REGISTRY[container_id]
-
-        print(f"✅ Container {container_id[:12]} cleaned up successfully")
-
-    except subprocess.TimeoutExpired:
-        print(f"⚠️  Timeout cleaning container {container_id[:12]} - forcing removal")
-        subprocess.run([engine_name, "kill", container_id], check=False)
-        subprocess.run([engine_name, "rm", "-f", container_id], check=False)
-    except Exception as e:
-        print(f"⚠️  Error cleaning container {container_id[:12]}: {e}")
-
-
-def pre_container_setup():
-    """Pre-container setup with disk space check and cleanup."""
-    if not CLEANUP_AVAILABLE:
-        return
-
-    try:
-        # Check disk space
-        if DiskSpaceMonitor.check_disk_space_critical(threshold_pct=85.0):
-            print("🚨 Low disk space detected - performing cleanup before container setup")
-            cleanup_manager.pre_test_cleanup(aggressive=True)
-
-        # Report disk usage
-        print(DiskSpaceMonitor.report_disk_usage(prefix="🧹"))
-
-    except Exception as e:
-        print(f"⚠️  Pre-container setup warning: {e}")
-
-
-# --- Helper Class for Engine Abstraction ---
+# --- Helper Class for Local Engine Execution ---
 
 
 class ExecutionEngine:
-    """A helper class to abstract command execution."""
+    """A helper class for local command execution."""
 
-    def __init__(self, engine_type: str, container_id: str | None = None):
+    def __init__(self, engine_type: str = "local"):
+        if engine_type != "local":
+            raise ValueError(f"Only local engine is supported, got: {engine_type}")
         self.engine_type = engine_type
-        self.container_id = container_id
-        print(f"\n✅ Engine initialized: type={self.engine_type}, container_id={self.container_id}")
+        print(f"\n✅ Engine initialized: type={self.engine_type}")
 
-    def run(self, command: list[str], **kwargs: Any) -> subprocess.CompletedProcess:
-        """Runs a command in the selected engine."""
+    def run(self, command: list[str], **kwargs) -> subprocess.CompletedProcess:
+        """Runs a command using local execution."""
         # Extract check parameter, default to True
         check = kwargs.pop("check", True)
 
         # Extract timeout parameter, default to 600 seconds (10 minutes) for long operations
         timeout = kwargs.pop("timeout", 600)
 
-        # Common kwargs for all engines
+        # Common kwargs for local execution
         run_kwargs = {"text": True, "capture_output": True, "check": check, "timeout": timeout, **kwargs}
 
-        if self.engine_type == "local":
-            return subprocess.run(command, **run_kwargs)
-
-        # Assumes podman commands are compatible with docker for exec
-        if self.engine_type in ["docker", "podman"]:
-            if self.container_id is None:
-                raise ValueError(f"Container ID required for {self.engine_type} engine")
-
-            # For containerized engines, handle cwd by using sh -c with cd
-            if "cwd" in kwargs:
-                cwd = kwargs.pop("cwd")
-                # Remove cwd from run_kwargs since it doesn't apply to host execution of docker
-                if "cwd" in run_kwargs:
-                    del run_kwargs["cwd"]
-                # Properly quote command arguments for shell
-                import shlex
-
-                quoted_command = " ".join(shlex.quote(arg) for arg in command)
-                # Use sh -c to change directory before executing command
-                shell_command = f"cd {shlex.quote(cwd)} && {quoted_command}"
-                base_command = [
-                    self.engine_type,
-                    "exec",
-                    self.container_id,
-                    "sh",
-                    "-c",
-                    shell_command,
-                ]
-                return subprocess.run(base_command, **run_kwargs)
-            else:
-                base_command = [self.engine_type, "exec", self.container_id]
-                full_command = base_command + command
-                return subprocess.run(full_command, **run_kwargs)
-
-        raise ValueError(f"Unsupported engine type: {self.engine_type}")
+        return subprocess.run(command, **run_kwargs)
 
     def rxiv_command(self, *args, **kwargs) -> subprocess.CompletedProcess:
-        """Standardized rxiv command execution across engines."""
+        """Standardized rxiv command execution for local engine."""
         import sys
 
-        if self.engine_type == "local":
-            try:
-                # Try uv run first (modern approach)
-                cmd = ["uv", "run", "rxiv"] + list(args)
-                return self.run(cmd, **kwargs)
-            except (FileNotFoundError, subprocess.CalledProcessError):
-                # Fallback to python module
-                cmd = [sys.executable, "-m", "rxiv_maker.cli"] + list(args)
-                return self.run(cmd, **kwargs)
-        else:
-            # In containers, rxiv should be installed
-            cmd = ["rxiv"] + list(args)
+        try:
+            # Try uv run first (modern approach)
+            cmd = ["uv", "run", "rxiv"] + list(args)
+            return self.run(cmd, **kwargs)
+        except (FileNotFoundError, subprocess.CalledProcessError):
+            # Fallback to python module
+            cmd = [sys.executable, "-m", "rxiv_maker.cli"] + list(args)
             return self.run(cmd, **kwargs)
 
 
@@ -192,112 +64,26 @@ class ExecutionEngine:
 
 
 def pytest_addoption(parser):
-    """Adds the --engine command-line option to pytest."""
+    """Adds the --engine command-line option to pytest (local only)."""
     parser.addoption(
         "--engine",
         action="store",
         default="local",
-        help="Specify the execution engine: local, docker, podman",
+        help="Specify the execution engine: local (only supported option)",
     )
 
 
 @pytest.fixture(scope="session")
 def execution_engine(request):
     """
-    Enhanced session-scoped fixture with container reuse and improved cleanup.
+    Session-scoped fixture for local execution engine.
     """
     engine_name = request.config.getoption("--engine")
 
-    if engine_name == "local":
-        yield ExecutionEngine("local")
-        return
+    if engine_name != "local":
+        pytest.fail(f"Only local engine is supported. Docker/Podman engines are deprecated. Got: {engine_name}")
 
-    # --- Containerized Engines (Docker, Podman, etc.) ---
-    container_id = None
-    reused_container = False
-
-    try:
-        if engine_name in ["docker", "podman"]:
-            # Pre-setup cleanup and disk space check
-            pre_container_setup()
-
-            # Use the existing rxiv-maker base image from Docker Hub
-            docker_image = "henriqueslab/rxiv-maker-base:latest"
-
-            # Try to reuse existing container first
-            container_id = get_reusable_container(engine_name, docker_image)
-            if container_id:
-                reused_container = True
-                # Verify rxiv-maker is still installed
-                try:
-                    result = subprocess.run(
-                        [engine_name, "exec", container_id, "rxiv", "--version"],
-                        capture_output=True,
-                        text=True,
-                        timeout=10,
-                    )
-                    if result.returncode != 0:
-                        print("🔄 Upgrading rxiv-maker to latest version in reused container...")
-                        subprocess.run(
-                            [engine_name, "exec", container_id, "/usr/local/bin/upgrade-to-latest-rxiv.sh"], check=True
-                        )
-                except Exception:
-                    # Container not suitable, create new one
-                    container_id = None
-                    reused_container = False
-
-            # Create new container if no reusable one found
-            if not container_id:
-                print(f"\n🐳 Pulling {engine_name} image: {docker_image}")
-                subprocess.run([engine_name, "pull", docker_image], check=True)
-
-                # Run the container in detached mode with workspace mounted
-                result = subprocess.run(
-                    [
-                        engine_name,
-                        "run",
-                        "-d",
-                        "--rm",
-                        "-v",
-                        f"{Path.cwd()}:/workspace",
-                        "-w",
-                        "/workspace",
-                        docker_image,
-                        "sleep",
-                        "infinity",
-                    ],
-                    check=True,
-                    text=True,
-                    capture_output=True,
-                )
-                container_id = result.stdout.strip()
-                print(f"\n🚀 Started {engine_name} container: {container_id[:12]}")
-
-                # Install rxiv-maker in the container
-                print("\n📦 Installing rxiv-maker in container...")
-                subprocess.run(
-                    [
-                        engine_name,
-                        "exec",
-                        container_id,
-                        "pip",
-                        "install",
-                        "-e",
-                        "/workspace",
-                    ],
-                    check=True,
-                )
-
-                # Register container for tracking
-                register_container(container_id, engine_name, docker_image, "session")
-
-            yield ExecutionEngine(engine_name, container_id)
-        else:
-            pytest.fail(f"Unsupported engine: {engine_name}")
-
-    finally:
-        if container_id and not reused_container:
-            cleanup_container(container_id, engine_name)
+    yield ExecutionEngine("local")
 
 
 # --- Optimized Temporary Directory Fixtures ---
@@ -455,13 +241,13 @@ def pytest_collection_modifyitems(config: pytest.Config, items: list[pytest.Item
         ):
             item.add_marker(requires_latex)
 
-        # Docker dependency detection
+        # Docker dependency detection (deprecated - for cleanup identification)
         if "docker" in test_name_lower or "container" in test_name_lower or "docker_engine" in test_file:
-            item.add_marker(requires_docker)
+            item.add_marker(pytest.mark.skip(reason="Docker engine deprecated - use local execution only"))
 
-        # Podman dependency detection
+        # Podman dependency detection (deprecated - for cleanup identification)
         if "podman" in test_name_lower or "podman_engine" in test_file:
-            item.add_marker(requires_podman)
+            item.add_marker(pytest.mark.skip(reason="Podman engine deprecated - use local execution only"))
 
         # R dependency detection
         if "r_" in test_name_lower or "_r_" in test_name_lower or "test_r" in test_name_lower:
@@ -469,7 +255,6 @@ def pytest_collection_modifyitems(config: pytest.Config, items: list[pytest.Item
 
         # Heavier or brittle unit tests to exclude by default in CI
         heavy_unit_files = {
-            "tests/unit/test_docker_engine_mode.py",
             "tests/unit/test_platform_detector.py",
             "tests/unit/test_figure_generator.py",
             "tests/unit/test_github_actions_integration.py",
@@ -618,29 +403,23 @@ def check_r_available():
         return False
 
 
+# Docker and Podman availability checks removed - engines deprecated
+# These functions kept as no-op for backward compatibility
 def check_docker_available():
-    """Check if Docker is available in the system."""
-    try:
-        result = subprocess.run(["docker", "--version"], capture_output=True, text=True)
-        return result.returncode == 0
-    except (FileNotFoundError, OSError):
-        return False
+    """Docker availability check (deprecated - always returns False)."""
+    return False
 
 
 def check_podman_available():
-    """Check if Podman is available in the system."""
-    try:
-        result = subprocess.run(["podman", "--version"], capture_output=True, text=True)
-        return result.returncode == 0
-    except (FileNotFoundError, OSError):
-        return False
+    """Podman availability check (deprecated - always returns False)."""
+    return False
 
 
 # Markers for conditional test execution
 requires_latex = pytest.mark.skipif(not check_latex_available(), reason="LaTeX not available")
 requires_r = pytest.mark.skipif(not check_r_available(), reason="R not available")
-requires_docker = pytest.mark.skipif(not check_docker_available(), reason="Docker not available")
-requires_podman = pytest.mark.skipif(not check_podman_available(), reason="Podman not available")
+requires_docker = pytest.mark.skip(reason="Docker engine deprecated - use local execution")
+requires_podman = pytest.mark.skip(reason="Podman engine deprecated - use local execution")
 
 # Test category markers (don't skip, just mark for selection)
 pytest.mark.unit = pytest.mark.unit
@@ -754,136 +533,25 @@ Missing proper structure.
 
 @pytest.fixture(scope="class")
 def class_execution_engine(request):
-    """Enhanced class-scoped execution engine with container reuse and optimized cleanup."""
+    """Class-scoped fixture for local execution engine."""
     engine_name = request.config.getoption("--engine")
 
-    if engine_name == "local":
-        yield ExecutionEngine("local")
-        return
+    if engine_name != "local":
+        pytest.fail(f"Only local engine is supported. Docker/Podman engines are deprecated. Got: {engine_name}")
 
-    # --- Containerized Engines (Docker, Podman, etc.) ---
-    container_id = None
-    reused_container = False
-
-    try:
-        if engine_name in ["docker", "podman"]:
-            # Pre-setup cleanup and disk space check
-            pre_container_setup()
-
-            # Use the existing rxiv-maker base image from Docker Hub
-            docker_image = "henriqueslab/rxiv-maker-base:latest"
-
-            # Try to reuse existing container first
-            container_id = get_reusable_container(engine_name, docker_image)
-            if container_id:
-                reused_container = True
-                print(f"✅ Reusing container: {container_id[:12]}")
-                # Verify rxiv-maker is still installed
-                try:
-                    result = subprocess.run(
-                        [engine_name, "exec", container_id, "rxiv", "--version"],
-                        capture_output=True,
-                        text=True,
-                        timeout=10,
-                    )
-                    if result.returncode != 0:
-                        print("🔄 Upgrading rxiv-maker to latest version in reused container...")
-                        subprocess.run(
-                            [engine_name, "exec", container_id, "/usr/local/bin/upgrade-to-latest-rxiv.sh"], check=True
-                        )
-                        print("✅ rxiv-maker upgraded to latest version successfully")
-                except Exception:
-                    # Container not suitable, create new one
-                    container_id = None
-                    reused_container = False
-
-            # Create new container if no reusable one found
-            if not container_id:
-                print(f"\n🐳 Pulling {engine_name} image: {docker_image}")
-                subprocess.run([engine_name, "pull", docker_image], check=True)
-
-                # Run the container in detached mode with workspace mounted
-                result = subprocess.run(
-                    [
-                        engine_name,
-                        "run",
-                        "-d",
-                        "-v",
-                        f"{Path.cwd()}:/workspace",
-                        "--workdir",
-                        "/workspace",
-                        docker_image,
-                        "sleep",
-                        "3600",  # Keep container alive for 1 hour
-                    ],
-                    capture_output=True,
-                    text=True,
-                    check=True,
-                )
-                container_id = result.stdout.strip()
-                print(f"✅ Container started: {container_id[:12]}")
-
-                # Install rxiv-maker in the container
-                print("📦 Installing rxiv-maker in container...")
-                subprocess.run(
-                    [
-                        engine_name,
-                        "exec",
-                        container_id,
-                        "pip",
-                        "install",
-                        "-e",
-                        "/workspace",
-                    ],
-                    check=True,
-                )
-                print("✅ rxiv-maker installed successfully")
-
-                # Register container for tracking
-                register_container(container_id, engine_name, docker_image, "class")
-
-            yield ExecutionEngine(engine_name, container_id)
-
-        else:
-            raise ValueError(f"Unsupported engine: {engine_name}")
-
-    finally:
-        # Clean up container only if we created it (not reused)
-        if container_id and not reused_container:
-            cleanup_container(container_id, engine_name)
+    yield ExecutionEngine("local")
 
 
 @pytest.fixture(autouse=True)
 def test_isolation():
-    """Enhanced test isolation with container and disk space awareness."""
-    # Pre-test setup
-    if CLEANUP_AVAILABLE:
-        try:
-            # Check for critical disk space before test
-            if DiskSpaceMonitor.check_disk_space_critical(threshold_pct=90.0):
-                print("🚨 Critical disk space - emergency cleanup before test")
-                cleanup_manager.emergency_cleanup()
-        except Exception as e:
-            print(f"⚠️  Pre-test disk check warning: {e}")
-
+    """Basic test isolation - optimized for performance."""
     yield
 
-    # Post-test cleanup
-    import gc
-
+    # Lightweight post-test cleanup
     # Clear any lingering environment variables
     test_env_vars = [var for var in os.environ if var.startswith("RXIV_TEST_")]
     for var in test_env_vars:
         os.environ.pop(var, None)
-
-    # Container-aware cleanup
-    if CLEANUP_AVAILABLE:
-        try:
-            # Light cleanup if disk space getting low
-            if DiskSpaceMonitor.check_disk_space_critical(threshold_pct=85.0):
-                cleanup_manager.post_test_cleanup(aggressive=False)
-        except Exception:
-            pass  # Don't fail tests due to cleanup issues
 
     # Force garbage collection
     gc.collect()
