+++ tests/regression/test_google_colab_integration.py
@@ -1,0 +1,86 @@
+"""Tests for Google Colab integration regression issues.
+
+This module contains regression tests for Google Colab specific integration
+problems identified by Guillaume, focusing on environment detection, path
+handling, and timeout management.
+
+Key issues tested:
+- Google Colab environment detection
+- Colab-specific path handling
+- Timeout handling for operations in Google Colab
+- Colab session compatibility
+"""
+
+import os
+import tempfile
+from pathlib import Path
+from unittest.mock import patch
+
+import pytest
+
+
+class TestGoogleColabIntegration:
+    """Test Google Colab specific integration issues."""
+
+    def test_colab_environment_detection(self):
+        """Test proper detection of Google Colab environment."""
+        # Test normal environment
+        assert not self._is_google_colab()
+
+        # Test with simulated Colab environment
+        with patch.dict(os.environ, {"COLAB_GPU": "0"}):
+            # Would be True if we had proper Colab detection
+            pass  # Placeholder for actual implementation
+
+    def test_colab_path_handling(self):
+        """Test path handling specific to Google Colab environment."""
+        with tempfile.TemporaryDirectory() as temp_dir:
+            # Simulate Colab-style paths
+            colab_content_path = Path(temp_dir) / "content"
+            colab_content_path.mkdir()
+
+            manuscript_dir = colab_content_path / "manuscript"
+            manuscript_dir.mkdir()
+            (manuscript_dir / "01_MAIN.md").write_text("# Colab Test")
+
+            # Test that paths are resolved correctly in Colab-like environment
+            from rxiv_maker.utils import find_manuscript_md
+
+            found_file = find_manuscript_md(manuscript_dir)
+            assert found_file is not None
+            assert "content" in str(found_file.parent)
+
+    def test_colab_timeout_handling(self):
+        """Test timeout handling for operations in Google Colab."""
+        # Colab sessions can timeout, so operations should be robust
+        with patch("subprocess.run") as mock_run:
+            # Simulate timeout
+            mock_run.side_effect = TimeoutError("Operation timed out")
+
+            from rxiv_maker.engines.operations.generate_figures import FigureGenerator
+
+            with tempfile.TemporaryDirectory() as temp_dir:
+                generator = FigureGenerator(figures_dir=temp_dir, output_dir=temp_dir, engine="local")
+
+                # Should handle timeout gracefully
+                try:
+                    generator.generate_all_figures()
+                except TimeoutError:
+                    pytest.fail("Timeout should be handled gracefully")
+
+    def _is_google_colab(self) -> bool:
+        """Check if running in Google Colab environment."""
+        try:
+            # Common ways to detect Colab
+            import google.colab  # noqa: F401
+
+            return True
+        except ImportError:
+            pass
+
+        # Check environment variables
+        return "COLAB_GPU" in os.environ or "COLAB_TPU_ADDR" in os.environ
+
+
+if __name__ == "__main__":
+    pytest.main([__file__])
