--- tests/regression/test_guillaume_issues.py
@@ -1,1734 +1,0 @@
-"""Tests for issues raised by Guillaume.
-
-This module contains regression tests for specific issues identified by Guillaume:
-- Issue #96: CLI path resolution problems
-- Issue #97: Google Colab argument parsing issues
-- PR #98: Widget authors being cleared when adding affiliations
-"""
-
-import os
-import tempfile
-from pathlib import Path
-from unittest.mock import Mock, patch
-
-import pytest
-
-
-class TestCLIArgumentParsing:
-    """Test CLI argument parsing issues (Issue #97)."""
-
-    def test_clean_command_with_unexpected_argument(self):
-        """Test that clean command properly handles unexpected arguments.
-
-        This tests the specific error from Issue #97:
-        'Error: Got unexpected extra argument (paper)'
-        """
-        from click.testing import CliRunner
-
-        from rxiv_maker.cli.main import main
-
-        runner = CliRunner()
-
-        # Test the problematic command that was failing in Google Colab
-        with tempfile.TemporaryDirectory() as temp_dir:
-            # Create a proper manuscript structure
-            manuscript_dir = Path(temp_dir) / "manuscript"
-            manuscript_dir.mkdir(parents=True)
-            (manuscript_dir / "01_MAIN.md").write_text("# Test")
-
-            # This should not cause "unexpected extra argument" error
-            result = runner.invoke(main, ["clean", str(manuscript_dir)], catch_exceptions=False)
-
-            # Should succeed or fail with a different error (not argument parsing)
-            assert "Got unexpected extra argument" not in result.output
-
-    def test_clean_command_argument_validation(self):
-        """Test clean command argument validation."""
-        from click.testing import CliRunner
-
-        from rxiv_maker.cli.main import main
-
-        runner = CliRunner()
-
-        # Test with invalid argument that should be caught properly
-        result = runner.invoke(main, ["clean", "--invalid-option"], catch_exceptions=True)
-
-        # Should give a helpful error message, not crash
-        assert result.exit_code != 0
-        assert "invalid-option" in result.output.lower() or "unknown option" in result.output.lower()
-
-    def test_pdf_command_argument_parsing(self):
-        """Test PDF command argument parsing for Google Colab compatibility."""
-        from click.testing import CliRunner
-
-        from rxiv_maker.cli.main import main
-
-        runner = CliRunner()
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            manuscript_dir = Path(temp_dir) / "manuscript"
-            manuscript_dir.mkdir(parents=True)
-            (manuscript_dir / "01_MAIN.md").write_text("# Test")
-
-            # Test the command that was run in Google Colab
-            result = runner.invoke(main, ["pdf", str(manuscript_dir)], catch_exceptions=True)
-
-            # Should not fail due to argument parsing
-            assert "Got unexpected extra argument" not in result.output
-
-
-class TestPathResolution:
-    """Test path resolution issues (Issue #96)."""
-
-    def test_manuscript_file_lookup_in_correct_directory(self):
-        """Test that manuscript files are looked up in the correct directory.
-
-        This addresses the issue where it was looking for 01_MAIN.md
-        in the parent folder instead of the manuscript folder.
-        """
-        from rxiv_maker.utils import find_manuscript_md
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            # Create manuscript structure
-            manuscript_dir = Path(temp_dir) / "test_manuscript"
-            manuscript_dir.mkdir(parents=True)
-            main_file = manuscript_dir / "01_MAIN.md"
-            main_file.write_text("# Test Manuscript")
-
-            # Should find the file in the manuscript directory
-            found_file = find_manuscript_md(manuscript_dir)
-            assert found_file is not None
-            assert found_file.name == "01_MAIN.md"
-            assert found_file.parent == manuscript_dir
-
-    def test_manuscript_file_lookup_with_environment_variable(self):
-        """Test manuscript lookup respects MANUSCRIPT_PATH environment variable."""
-        from rxiv_maker.utils import find_manuscript_md
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            manuscript_dir = Path(temp_dir) / "env_manuscript"
-            manuscript_dir.mkdir(parents=True)
-            main_file = manuscript_dir / "01_MAIN.md"
-            main_file.write_text("# Test Manuscript")
-
-            # Test with environment variable set
-            with patch.dict(os.environ, {"MANUSCRIPT_PATH": str(manuscript_dir)}):
-                found_file = find_manuscript_md()
-                assert found_file is not None
-                assert found_file.parent == manuscript_dir
-
-    def test_figure_path_resolution(self):
-        """Test figure path resolution and display consistency.
-
-        This addresses issues with figure path display from Issue #96.
-        """
-        from rxiv_maker.engines.operations.generate_figures import FigureGenerator
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            manuscript_dir = Path(temp_dir)
-            figures_dir = manuscript_dir / "FIGURES"
-            figures_dir.mkdir(parents=True)
-
-            # Create a test figure script
-            test_script = figures_dir / "Figure__test.py"
-            test_script.write_text("""
-import matplotlib.pyplot as plt
-plt.figure()
-plt.plot([1, 2, 3], [1, 4, 9])
-plt.savefig('Figure__test.png')
-plt.close()
-""")
-
-            FigureGenerator(figures_dir=str(figures_dir), output_dir=str(figures_dir), engine="local")
-
-            # Should properly resolve paths without looking in parent directories
-            python_files = list(figures_dir.glob("*.py"))
-            assert len(python_files) > 0
-            assert any(f.name == "Figure__test.py" for f in python_files)
-
-    def test_working_directory_independence(self):
-        """Test that operations work regardless of current working directory."""
-        from click.testing import CliRunner
-
-        from rxiv_maker.cli.main import main
-
-        runner = CliRunner()
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            # Create manuscript in subdirectory
-            manuscript_dir = Path(temp_dir) / "project" / "manuscript"
-            manuscript_dir.mkdir(parents=True)
-            (manuscript_dir / "01_MAIN.md").write_text("# Test")
-
-            # Create another directory to run from
-            run_dir = Path(temp_dir) / "other_directory"
-            run_dir.mkdir()
-
-            # Change to different directory and run command
-            original_cwd = os.getcwd()
-            try:
-                os.chdir(run_dir)
-                result = runner.invoke(main, ["validate", str(manuscript_dir)], catch_exceptions=True)
-
-                # Should work correctly even when run from different directory
-                assert "not found in" not in result.output
-                # May have validation errors, but shouldn't have path resolution errors
-
-            finally:
-                os.chdir(original_cwd)
-
-
-class TestWidgetAuthorBehavior:
-    """Test widget behavior for author/affiliation handling (PR #98)."""
-
-    @pytest.fixture
-    def mock_widget_environment(self):
-        """Set up mock widget environment for testing."""
-        # Mock IPython/Jupyter environment
-        mock_display = Mock()
-        mock_widget = Mock()
-
-        # Create comprehensive mocks for the entire IPython ecosystem
-        mock_ipython_display = Mock()
-        mock_ipython_display.display = mock_display
-        mock_ipython_display.clear_output = Mock()
-
-        mock_ipywidgets = Mock()
-        mock_ipywidgets.Widget = mock_widget
-
-        # Mock the modules in sys.modules to avoid import errors
-        with patch.dict(
-            "sys.modules", {"IPython": Mock(), "IPython.display": mock_ipython_display, "ipywidgets": mock_ipywidgets}
-        ):
-            yield {"display": mock_display, "widget": mock_widget}
-
-    def test_author_widget_preservation_on_affiliation_add(self, mock_widget_environment):
-        """Test that authors are not cleared when adding affiliations.
-
-        This addresses the specific issue in PR #98 where authors were
-        being cleared every time an affiliation was added.
-        """
-        # This test would need to be implemented once we have access to the widget code
-        # For now, we'll create a placeholder that demonstrates the expected behavior
-
-        # Simulate widget state
-        authors = ["John Doe", "Jane Smith"]
-        affiliations = ["University A"]
-
-        # Simulate adding an affiliation
-        new_affiliation = "University B"
-        affiliations.append(new_affiliation)
-
-        # Authors should remain unchanged
-        expected_authors = ["John Doe", "Jane Smith"]
-        assert authors == expected_authors
-
-        # But affiliations should be updated
-        expected_affiliations = ["University A", "University B"]
-        assert affiliations == expected_affiliations
-
-    def test_widget_state_consistency(self, mock_widget_environment):
-        """Test that widget state remains consistent during updates."""
-        # Placeholder for widget state consistency test
-        # This would test the actual widget behavior once the widget code is available
-
-        initial_state = {"authors": ["Author 1", "Author 2"], "affiliations": ["Affiliation 1"], "title": "Test Paper"}
-
-        # Simulate state update that should not affect other fields
-        updated_state = initial_state.copy()
-        updated_state["affiliations"].append("Affiliation 2")
-
-        # Other fields should remain unchanged
-        assert updated_state["authors"] == initial_state["authors"]
-        assert updated_state["title"] == initial_state["title"]
-        assert len(updated_state["affiliations"]) == 2
-
-
-class TestGoogleColabIntegration:
-    """Test Google Colab specific integration issues."""
-
-    def test_colab_environment_detection(self):
-        """Test proper detection of Google Colab environment."""
-        # Test normal environment
-        assert not self._is_google_colab()
-
-        # Test with simulated Colab environment
-        with patch.dict(os.environ, {"COLAB_GPU": "0"}):
-            # Would be True if we had proper Colab detection
-            pass  # Placeholder for actual implementation
-
-    def test_colab_path_handling(self):
-        """Test path handling specific to Google Colab environment."""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            # Simulate Colab-style paths
-            colab_content_path = Path(temp_dir) / "content"
-            colab_content_path.mkdir()
-
-            manuscript_dir = colab_content_path / "manuscript"
-            manuscript_dir.mkdir()
-            (manuscript_dir / "01_MAIN.md").write_text("# Colab Test")
-
-            # Test that paths are resolved correctly in Colab-like environment
-            from rxiv_maker.utils import find_manuscript_md
-
-            found_file = find_manuscript_md(manuscript_dir)
-            assert found_file is not None
-            assert "content" in str(found_file.parent)
-
-    def test_colab_timeout_handling(self):
-        """Test timeout handling for operations in Google Colab."""
-        # Colab sessions can timeout, so operations should be robust
-        with patch("subprocess.run") as mock_run:
-            # Simulate timeout
-            mock_run.side_effect = TimeoutError("Operation timed out")
-
-            from rxiv_maker.engines.operations.generate_figures import FigureGenerator
-
-            with tempfile.TemporaryDirectory() as temp_dir:
-                generator = FigureGenerator(figures_dir=temp_dir, output_dir=temp_dir, engine="local")
-
-                # Should handle timeout gracefully
-                try:
-                    generator.generate_all_figures()
-                except TimeoutError:
-                    pytest.fail("Timeout should be handled gracefully")
-
-    def _is_google_colab(self) -> bool:
-        """Check if running in Google Colab environment."""
-        try:
-            # Common ways to detect Colab
-            import google.colab  # noqa: F401
-
-            return True
-        except ImportError:
-            pass
-
-        # Check environment variables
-        return "COLAB_GPU" in os.environ or "COLAB_TPU_ADDR" in os.environ
-
-
-class TestErrorMessageQuality:
-    """Test that error messages are helpful for debugging Guillaume's issues."""
-
-    def test_path_not_found_error_messages(self):
-        """Test that path not found errors provide helpful information."""
-        from rxiv_maker.utils import find_manuscript_md
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            empty_dir = Path(temp_dir) / "empty"
-            empty_dir.mkdir()
-
-            # Should raise a helpful error message when file not found
-            with pytest.raises(FileNotFoundError) as exc_info:
-                find_manuscript_md(empty_dir)
-
-            # Error message should be helpful and mention the directory
-            error_msg = str(exc_info.value)
-            assert "01_MAIN.md not found" in error_msg
-            assert str(empty_dir) in error_msg
-
-    def test_cli_help_messages(self):
-        """Test that CLI help messages are clear and helpful."""
-        from click.testing import CliRunner
-
-        from rxiv_maker.cli.main import main
-
-        runner = CliRunner()
-
-        # Test main help
-        result = runner.invoke(main, ["--help"])
-        assert result.exit_code == 0
-        assert "manuscript" in result.output.lower()
-
-        # Test clean command help
-        result = runner.invoke(main, ["clean", "--help"])
-        assert result.exit_code == 0
-        assert "clean" in result.output.lower()
-
-    def test_validation_error_clarity(self):
-        """Test that validation errors are clear and actionable."""
-        from click.testing import CliRunner
-
-        from rxiv_maker.cli.main import main
-
-        runner = CliRunner()
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            # Create incomplete manuscript
-            manuscript_dir = Path(temp_dir) / "incomplete"
-            manuscript_dir.mkdir()
-            # Missing 01_MAIN.md file
-
-            result = runner.invoke(main, ["validate", str(manuscript_dir)], catch_exceptions=True)
-
-            # Should provide clear error about missing file
-            assert "01_MAIN.md" in result.output or "main" in result.output.lower()
-
-
-class TestWidgetInteractionsWithPlaywright:
-    """Test widget interactions using Playwright for Google Colab compatibility.
-
-    These tests address PR #98: authors being cleared when adding affiliations.
-    """
-
-    @pytest.fixture
-    def browser_context(self):
-        """Set up browser context for widget testing."""
-        pytest.importorskip("playwright")
-        from playwright.sync_api import sync_playwright
-
-        with sync_playwright() as p:
-            browser = p.chromium.launch(headless=True)
-            context = browser.new_context()
-            yield context
-            browser.close()
-
-    def test_colab_notebook_widget_loading(self, browser_context):
-        """Test that widgets load properly in a Colab-like environment."""
-        page = browser_context.new_page()
-
-        # Create a minimal HTML page that simulates Colab notebook interface
-        html_content = """
-        <!DOCTYPE html>
-        <html>
-        <head>
-            <title>Test Colab Notebook</title>
-            <script src="https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.6/require.min.js"></script>
-            <style>
-                .widget-container { padding: 10px; margin: 10px; border: 1px solid #ccc; }
-                .author-widget { background: #f5f5f5; }
-                .affiliation-widget { background: #e5f5e5; }
-                .button { padding: 5px 10px; margin: 5px; cursor: pointer; }
-                .text-input { padding: 5px; margin: 5px; width: 200px; }
-            </style>
-        </head>
-        <body>
-            <div id="notebook-container">
-                <h1>Test Notebook for rxiv-maker Widget</h1>
-
-                <!-- Simulate the author/affiliation widget -->
-                <div class="widget-container author-widget">
-                    <h3>Authors</h3>
-                    <div id="authors-list">
-                        <div class="author-entry">
-                            <input type="text" class="text-input author-name" placeholder="Author name" value="John Doe">
-                            <button class="button remove-author">Remove</button>
-                        </div>
-                    </div>
-                    <button class="button" id="add-author">Add Author</button>
-                </div>
-
-                <div class="widget-container affiliation-widget">
-                    <h3>Affiliations</h3>
-                    <div id="affiliations-list">
-                        <div class="affiliation-entry">
-                            <input type="text" class="text-input affiliation-name" placeholder="Affiliation" value="University A">
-                            <button class="button remove-affiliation">Remove</button>
-                        </div>
-                    </div>
-                    <button class="button" id="add-affiliation">Add Affiliation</button>
-                </div>
-            </div>
-
-            <script>
-                // Simulate the widget behavior that was causing issues
-                document.getElementById('add-affiliation').addEventListener('click', function() {
-                    // This simulates the bug where authors were cleared when adding affiliations
-                    var affiliationsList = document.getElementById('affiliations-list');
-                    var newAffiliation = document.createElement('div');
-                    newAffiliation.className = 'affiliation-entry';
-                    newAffiliation.innerHTML = '<input type="text" class="text-input affiliation-name" placeholder="New affiliation">' +
-                                              '<button class="button remove-affiliation">Remove</button>';
-                    affiliationsList.appendChild(newAffiliation);
-
-                    // The bug: DO NOT clear authors when adding affiliations
-                    // This is what the original bug was doing - we test that it doesn't happen
-                    console.log('Added affiliation without clearing authors');
-                });
-
-                document.getElementById('add-author').addEventListener('click', function() {
-                    var authorsList = document.getElementById('authors-list');
-                    var newAuthor = document.createElement('div');
-                    newAuthor.className = 'author-entry';
-                    newAuthor.innerHTML = '<input type="text" class="text-input author-name" placeholder="New author">' +
-                                         '<button class="button remove-author">Remove</button>';
-                    authorsList.appendChild(newAuthor);
-                });
-
-                // Add event delegation for remove buttons
-                document.addEventListener('click', function(e) {
-                    if (e.target.classList.contains('remove-author')) {
-                        e.target.parentElement.remove();
-                    } else if (e.target.classList.contains('remove-affiliation')) {
-                        e.target.parentElement.remove();
-                    }
-                });
-            </script>
-        </body>
-        </html>
-        """
-
-        # Load the test page
-        page.set_content(html_content)
-
-        # Wait for the page to load completely
-        page.wait_for_selector("#add-author")
-        page.wait_for_selector("#add-affiliation")
-
-        # Get initial author count
-        initial_authors = page.query_selector_all(".author-entry")
-        assert len(initial_authors) == 1
-
-        # Get initial author value
-        initial_author_name = page.query_selector(".author-name").input_value()
-        assert initial_author_name == "John Doe"
-
-        # Add a new affiliation (this was causing the bug)
-        page.click("#add-affiliation")
-
-        # Verify that authors are NOT cleared (this is the fix)
-        authors_after_affiliation = page.query_selector_all(".author-entry")
-        assert len(authors_after_affiliation) == 1  # Should still have the original author
-
-        # Verify the original author name is still there
-        author_name_after = page.query_selector(".author-name").input_value()
-        assert author_name_after == "John Doe"  # Should not be cleared
-
-        # Verify the new affiliation was added
-        affiliations = page.query_selector_all(".affiliation-entry")
-        assert len(affiliations) == 2  # Original + newly added
-
-    def test_widget_state_persistence_across_interactions(self, browser_context):
-        """Test that widget state persists across multiple interactions."""
-        page = browser_context.new_page()
-
-        # Minimal widget testing page
-        html_content = """
-        <!DOCTYPE html>
-        <html>
-        <head>
-            <title>Widget State Test</title>
-            <style>
-                .widget { padding: 10px; margin: 10px; border: 1px solid #ddd; }
-                .input-field { padding: 5px; margin: 5px; width: 200px; }
-                .button { padding: 5px 10px; margin: 5px; cursor: pointer; }
-            </style>
-        </head>
-        <body>
-            <div class="widget">
-                <input type="text" id="author1" class="input-field" placeholder="Author 1" value="">
-                <input type="text" id="author2" class="input-field" placeholder="Author 2" value="">
-                <input type="text" id="affiliation1" class="input-field" placeholder="Affiliation 1" value="">
-                <button id="simulate-interaction" class="button">Simulate Interaction</button>
-                <div id="state-display"></div>
-            </div>
-
-            <script>
-                document.getElementById('simulate-interaction').addEventListener('click', function() {
-                    // This simulates the kind of interaction that was causing state loss
-                    var stateDisplay = document.getElementById('state-display');
-                    var author1 = document.getElementById('author1').value;
-                    var author2 = document.getElementById('author2').value;
-                    var affiliation1 = document.getElementById('affiliation1').value;
-
-                    stateDisplay.innerHTML = 'State preserved: ' +
-                        'Author1=' + author1 + ', Author2=' + author2 + ', Affiliation1=' + affiliation1;
-                });
-            </script>
-        </body>
-        </html>
-        """
-
-        page.set_content(html_content)
-        page.wait_for_selector("#author1")
-
-        # Fill in some data
-        page.fill("#author1", "Alice Smith")
-        page.fill("#author2", "Bob Jones")
-        page.fill("#affiliation1", "MIT")
-
-        # Trigger interaction that might cause state loss
-        page.click("#simulate-interaction")
-
-        # Verify state is preserved
-        state_text = page.text_content("#state-display")
-        assert "Alice Smith" in state_text
-        assert "Bob Jones" in state_text
-        assert "MIT" in state_text
-
-        # Verify inputs still have their values
-        assert page.input_value("#author1") == "Alice Smith"
-        assert page.input_value("#author2") == "Bob Jones"
-        assert page.input_value("#affiliation1") == "MIT"
-
-    def test_colab_ipywidgets_compatibility(self, browser_context):
-        """Test compatibility with IPython widgets environment."""
-        page = browser_context.new_page()
-
-        # Simulate the IPython widgets environment
-        html_content = """
-        <!DOCTYPE html>
-        <html>
-        <head>
-            <title>IPython Widgets Test</title>
-            <script>
-                // Mock IPython environment
-                window.IPython = {
-                    notebook: {
-                        kernel: {
-                            execute: function(code) {
-                                console.log('Executing:', code);
-                                return { then: function(callback) { callback(); } };
-                            }
-                        }
-                    }
-                };
-
-                // Mock Jupyter widgets
-                window.jupyter = {
-                    widgets: {
-                        output: {
-                            clear_output: function() {
-                                console.log('Clearing output');
-                            }
-                        }
-                    }
-                };
-            </script>
-            <style>
-                .jupyter-widgets { padding: 10px; border: 1px solid #ccc; }
-                .widget-text { padding: 5px; margin: 5px; }
-                .widget-button { padding: 5px 10px; margin: 5px; }
-            </style>
-        </head>
-        <body>
-            <div class="jupyter-widgets">
-                <h3>rxiv-maker Widget Test</h3>
-                <div class="widget-text">
-                    <label>Manuscript Title:</label>
-                    <input type="text" id="manuscript-title" value="My Research Paper">
-                </div>
-                <div class="widget-text">
-                    <label>Authors:</label>
-                    <textarea id="authors-textarea" rows="3">Author 1, Author 2</textarea>
-                </div>
-                <button class="widget-button" id="update-metadata">Update Metadata</button>
-                <div id="result"></div>
-            </div>
-
-            <script>
-                document.getElementById('update-metadata').addEventListener('click', function() {
-                    var title = document.getElementById('manuscript-title').value;
-                    var authors = document.getElementById('authors-textarea').value;
-
-                    // Simulate the widget updating metadata
-                    document.getElementById('result').innerHTML =
-                        'Updated: Title="' + title + '", Authors="' + authors + '"';
-
-                    // This is where the bug would manifest - losing data during updates
-                    console.log('Metadata updated without data loss');
-                });
-            </script>
-        </body>
-        </html>
-        """
-
-        page.set_content(html_content)
-        page.wait_for_selector("#manuscript-title")
-
-        # Verify initial state
-        assert page.input_value("#manuscript-title") == "My Research Paper"
-        assert "Author 1, Author 2" in page.input_value("#authors-textarea")
-
-        # Modify data
-        page.fill("#manuscript-title", "Updated Research Paper")
-        page.fill("#authors-textarea", "Alice Smith, Bob Jones, Carol White")
-
-        # Trigger update (this is where the bug would occur)
-        page.click("#update-metadata")
-
-        # Verify data persistence after update
-        result_text = page.text_content("#result")
-        assert "Updated Research Paper" in result_text
-        assert "Alice Smith, Bob Jones, Carol White" in result_text
-
-        # Verify inputs still have the updated values
-        assert page.input_value("#manuscript-title") == "Updated Research Paper"
-        assert "Alice Smith, Bob Jones, Carol White" in page.input_value("#authors-textarea")
-
-    def test_colab_environment_variables_handling(self, browser_context):
-        """Test handling of Google Colab environment variables and paths."""
-        page = browser_context.new_page()
-
-        # Simulate Colab environment detection
-        html_content = """
-        <!DOCTYPE html>
-        <html>
-        <head>
-            <title>Colab Environment Test</title>
-        </head>
-        <body>
-            <div id="environment-info">
-                <h3>Environment Detection</h3>
-                <div id="colab-status">Unknown</div>
-                <div id="path-info"></div>
-            </div>
-
-            <script>
-                // Simulate environment detection logic
-                function detectColabEnvironment() {
-                    var isColab = window.location.hostname.includes('colab.research.google.com') ||
-                                 document.getElementById('site-name') !== null ||
-                                 navigator.userAgent.includes('Colab');
-
-                    document.getElementById('colab-status').textContent =
-                        isColab ? 'Google Colab Detected' : 'Local Environment';
-
-                    // Simulate path handling that was problematic in Guillaume's issues
-                    var paths = {
-                        working_dir: '/content',
-                        manuscript_dir: '/content/manuscript',
-                        figures_dir: '/content/manuscript/FIGURES'
-                    };
-
-                    document.getElementById('path-info').innerHTML =
-                        'Working Dir: ' + paths.working_dir + '<br>' +
-                        'Manuscript Dir: ' + paths.manuscript_dir + '<br>' +
-                        'Figures Dir: ' + paths.figures_dir;
-                }
-
-                detectColabEnvironment();
-            </script>
-        </body>
-        </html>
-        """
-
-        page.set_content(html_content)
-        page.wait_for_selector("#colab-status")
-
-        # Verify environment detection works
-        status_text = page.text_content("#colab-status")
-        assert "Environment" in status_text
-
-        # Verify path information is displayed
-        path_info = page.text_content("#path-info")
-        assert "/content" in path_info
-        assert "manuscript" in path_info.lower()
-
-
-class TestDiscordReportedIssues:
-    """Test specific issues reported by Guillaume in Discord messages."""
-
-    def test_latex_dependency_packages_included(self):
-        """Test that missing LaTeX packages (siunitx, ifsym) are included in dependencies."""
-        from rxiv_maker.install.dependency_handlers.latex import LaTeXHandler
-        from rxiv_maker.install.utils.logging import InstallLogger
-
-        logger = InstallLogger()
-        handler = LaTeXHandler(logger)
-        essential_packages = handler.get_essential_packages()
-
-        # Guillaume reported missing siunitx.sty and ifsym.sty
-        assert "siunitx" in essential_packages, "siunitx package should be in essential packages"
-        assert "ifsym" in essential_packages, "ifsym package should be in essential packages (Guillaume's issue)"
-
-    def test_debian_control_latex_dependencies(self):
-        """Test that Debian control file includes the LaTeX packages Guillaume needed."""
-
-        control_file_path = Path(__file__).parent.parent.parent / "packaging" / "debian" / "control"
-
-        if control_file_path.exists():
-            content = control_file_path.read_text()
-
-            # Guillaume needed these packages to fix siunitx.sty and ifsym.sty errors
-            assert "texlive-science" in content, "texlive-science should be in Debian dependencies"
-            assert "texlive-fonts-extra" in content, "texlive-fonts-extra should be in Debian dependencies"
-
-    def test_figure_panel_reference_spacing_fix(self):
-        """Test that figure panel references don't have unwanted spaces.
-
-        Guillaume reported: (@fig:Figure1 A) renders as (Fig. 1 A) instead of (Fig. 1A)
-        """
-        from rxiv_maker.converters.figure_processor import convert_figure_references_to_latex
-
-        # Test the specific case Guillaume reported
-        test_text = "As shown in (@fig:Figure1 A), the results indicate..."
-        result = convert_figure_references_to_latex(test_text)
-
-        # Should render as Fig. \ref{fig:Figure1}{}A (no space between ref and A)
-        assert "Fig. \\ref{fig:Figure1}{}A)" in result, (
-            f"Expected empty group {{}} spacing control between figure ref and panel letter, got: {result}"
-        )
-
-        # Should NOT have a space
-        assert "Fig. \\ref{fig:Figure1} A)" not in result, "Should not have space between figure ref and panel letter"
-
-        # Test supplementary figures too
-        test_text_sfig = "As shown in (@sfig:SupFig1 B), the analysis shows..."
-        result_sfig = convert_figure_references_to_latex(test_text_sfig)
-
-        assert "Fig. \\ref{sfig:SupFig1}{}B)" in result_sfig, (
-            "Supplementary figure panel refs should also have empty group spacing control"
-        )
-
-    def test_figure_ready_files_loading_fix(self):
-        """Test that ready figures load correctly without requiring subdirectory duplication.
-
-        Guillaume reported: need to have Fig1.png in both Figure/ and Figure/Fig1/Fig1.png
-        """
-        import os
-        import tempfile
-        from pathlib import Path
-
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        with tempfile.TemporaryDirectory() as tmpdir:
-            tmpdir_path = Path(tmpdir)
-            original_cwd = os.getcwd()
-            os.chdir(tmpdir)
-
-            try:
-                # Test Case 1: With ready file - should use direct path
-                figures_dir = tmpdir_path / "FIGURES"
-                figures_dir.mkdir()
-                ready_file = figures_dir / "Fig1.png"
-                ready_file.write_text("fake png content")
-
-                latex_result_with_ready = create_latex_figure_environment(
-                    path="FIGURES/Fig1.png", caption="Test figure caption", attributes={}
-                )
-
-                # Should use direct path, not subdirectory format
-                assert "Figures/Fig1.png" in latex_result_with_ready, "Should use ready file directly: Figures/Fig1.png"
-                assert "Figures/Fig1/Fig1.png" not in latex_result_with_ready, (
-                    "Should NOT use subdirectory format when ready file exists"
-                )
-
-                # Test Case 2: Without ready file - should use subdirectory format
-                ready_file.unlink()  # Remove the ready file
-
-                latex_result_without_ready = create_latex_figure_environment(
-                    path="FIGURES/Fig1.png", caption="Test figure caption", attributes={}
-                )
-
-                # Should use direct format (Guillaume's implementation)
-                assert "Figures/Fig1.png" in latex_result_without_ready, (
-                    "Should use direct format (Guillaume's implementation)"
-                )
-                # Guillaume's implementation uses direct format consistently
-
-            finally:
-                os.chdir(original_cwd)
-
-    def test_section_header_introduction_preservation(self):
-        """Test that ## Introduction stays as Introduction, not mapped to Main.
-
-        Guillaume reported: ## Introduction renders as Main in PDF instead of Introduction
-        """
-        from rxiv_maker.converters.section_processor import map_section_title_to_key
-
-        # Test the specific case Guillaume reported
-        result = map_section_title_to_key("Introduction")
-        assert result == "introduction", f"Introduction should map to 'introduction', not 'main'. Got: {result}"
-
-        # Test case variations
-        assert map_section_title_to_key("introduction") == "introduction"
-        assert map_section_title_to_key("INTRODUCTION") == "introduction"
-
-        # Test with content sections extraction
-        from rxiv_maker.converters.section_processor import extract_content_sections
-
-        # Create test content with Introduction section
-        test_markdown = """# Test Article
-
-## Introduction
-
-This is the introduction content.
-
-## Methods
-
-This is the methods content.
-"""
-
-        sections = extract_content_sections(test_markdown)
-
-        # Should have introduction as a separate key, not mapped to main
-        assert "introduction" in sections, "Should extract introduction section"
-        assert "This is the introduction content" in sections["introduction"], (
-            "Introduction content should be preserved"
-        )
-
-        # NEW: Test the actual template processing - this was the real issue!
-        from rxiv_maker.processors.template_processor import process_template_replacements
-
-        # Create minimal template content with the main section placeholder
-        template_content = """
-<PY-RPL:MAIN-SECTION>
-
-<PY-RPL:METHODS>
-"""
-
-        # Process template with introduction section
-        yaml_metadata = {}
-        result = process_template_replacements(template_content, yaml_metadata, test_markdown)
-
-        # Should create an Introduction section, not Main
-        assert "\\section*{Introduction}" in result, "Template should create Introduction section header, not Main"
-        assert "This is the introduction content" in result, "Introduction content should be in final template"
-        assert "\\section*{Main}" not in result, "Should NOT create Main section when Introduction exists"
-
-    def test_full_page_figure_positioning_fix(self):
-        """Test that full-page figures with textwidth don't break positioning.
-
-        Guillaume reported: tex_position="p" with width=textwidth creates 2-column layout instead of dedicated page
-        """
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test Guillaume's specific case: textwidth with position p should use regular figure environment
-        latex_result = create_latex_figure_environment(
-            path="FIGURES/Figure__workflow.svg",
-            caption="Test figure caption",
-            attributes={"width": "\\textwidth", "tex_position": "p", "id": "fig:workflow"},
-        )
-
-        # Should use figure* environment for dedicated page to span full width in two-column documents
-        assert "\\begin{figure*}[p]" in latex_result, (
-            "Dedicated page figures should use figure*[p] for full-width spanning"
-        )
-        assert "\\begin{figure}[p]" not in latex_result, (
-            "Should use figure*[p], not figure[p], for dedicated page positioning"
-        )
-
-        # Test comparison: textwidth without explicit position should use figure*
-        latex_result2 = create_latex_figure_environment(
-            path="FIGURES/Figure__workflow.svg",
-            caption="Test figure caption",
-            attributes={"width": "\\textwidth", "id": "fig:workflow"},  # No tex_position
-        )
-
-        # Should use figure* for 2-column spanning when no explicit position
-        assert "\\begin{figure*}" in latex_result2, (
-            "Full-width figures should use figure* for 2-column spanning by default"
-        )
-
-        # Test that other positioning is preserved with figure*
-        latex_result3 = create_latex_figure_environment(
-            path="FIGURES/Figure__workflow.svg",
-            caption="Test figure caption",
-            attributes={"width": "\\textwidth", "tex_position": "t", "id": "fig:workflow"},
-        )
-
-        # Should use figure* with user's positioning
-        assert "\\begin{figure*}[t]" in latex_result3, "Should respect user's tex_position when using figure*"
-
-    def test_latex_package_installation_verification(self):
-        """Test that LaTeX package installation verification works for Guillaume's packages."""
-        from rxiv_maker.install.dependency_handlers.latex import LaTeXHandler
-        from rxiv_maker.install.utils.logging import InstallLogger
-
-        logger = InstallLogger()
-        handler = LaTeXHandler(logger)
-
-        # Test that verification would work for the packages Guillaume needed
-        essential_packages = handler.get_essential_packages()
-
-        # These are the packages Guillaume had to install manually
-        required_packages = ["siunitx", "ifsym"]
-
-        for package in required_packages:
-            assert package in essential_packages, (
-                f"Package {package} should be in essential list for automatic installation"
-            )
-
-    def test_figure_reference_edge_cases(self):
-        """Test edge cases for figure references that might cause spacing issues."""
-        from rxiv_maker.converters.figure_processor import convert_figure_references_to_latex
-
-        # Test various panel reference formats
-        test_cases = [
-            ("(@fig:Figure1 A)", "Fig. \\ref{fig:Figure1}{}A)"),
-            ("(@fig:Figure1 B)", "Fig. \\ref{fig:Figure1}{}B)"),
-            ("(@fig:Figure1 C) and (@fig:Figure2 D)", "Fig. \\ref{fig:Figure1}{}C) and (Fig. \\ref{fig:Figure2}{}D)"),
-            ("@fig:Figure1 A shows", "Fig. \\ref{fig:Figure1}{}A shows"),  # Without parentheses
-            ("(@sfig:SupFig1 A)", "Fig. \\ref{sfig:SupFig1}{}A)"),  # Supplementary figures
-        ]
-
-        for input_text, expected_pattern in test_cases:
-            result = convert_figure_references_to_latex(input_text)
-            assert expected_pattern in result, (
-                f"Failed for input '{input_text}': expected '{expected_pattern}' in '{result}'"
-            )
-
-    def test_integration_all_guillaume_fixes_together(self):
-        """Integration test that all Guillaume's fixes work together."""
-        from rxiv_maker.converters.figure_processor import (
-            convert_figure_references_to_latex,
-            create_latex_figure_environment,
-        )
-        from rxiv_maker.converters.section_processor import map_section_title_to_key
-        from rxiv_maker.install.dependency_handlers.latex import LaTeXHandler
-        from rxiv_maker.install.utils.logging import InstallLogger
-
-        # Test all fixes work together
-        logger = InstallLogger()
-        latex_handler = LaTeXHandler(logger)
-
-        # 1. LaTeX dependencies should include Guillaume's packages
-        packages = latex_handler.get_essential_packages()
-        assert "siunitx" in packages and "ifsym" in packages, "Guillaume's required packages should be included"
-
-        # 2. Figure panel references should work correctly
-        panel_ref = convert_figure_references_to_latex("(@fig:Figure1 A)")
-        assert "Fig. \\ref{fig:Figure1}{}A)" in panel_ref, "Panel references should use empty group for spacing control"
-
-        # 3. Section mapping should preserve Introduction
-        section_key = map_section_title_to_key("Introduction")
-        assert section_key == "introduction", "Introduction should not be mapped to main"
-
-        # 4. Figure positioning should respect user preferences
-        figure_latex = create_latex_figure_environment(
-            path="FIGURES/test.svg", caption="Test caption", attributes={"width": "\\textwidth", "tex_position": "t"}
-        )
-        assert "\\begin{figure*}[t]" in figure_latex, "Should respect user's positioning preference"
-
-        print("✅ All Guillaume's fixes are working together correctly")
-
-    def test_dedicated_page_figure_caption_width(self):
-        """Test that dedicated page figures have full-width captions.
-
-        Guillaume reported: Dedicated page figure captions were too narrow, not spanning full page width
-        """
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test dedicated page figure with textwidth
-        result = create_latex_figure_environment(
-            path="FIGURES/test.png",
-            caption="Test caption for dedicated page figure",
-            attributes={"tex_position": "p", "width": "\\textwidth", "id": "fig:test"},
-        )
-
-        # Should use figure*[p] for dedicated page to span full width
-        assert "\\begin{figure*}[p]" in result, (
-            "Dedicated page figures should use figure*[p] to span full width in two-column documents"
-        )
-
-        # Should have width=\linewidth in captionsetup for full-width caption
-        assert "\\captionsetup{width=0.95\\textwidth" in result, (
-            "Dedicated page figures should use width=0.95\\textwidth (Guillaume's implementation)"
-        )
-        # Note: justification=justified only added for longer captions (>150 chars)
-
-        # Should use figure*[p] for proper dedicated page control
-        assert "\\begin{figure*}[p]" in result, "Should use figure*[p] for dedicated page placement"
-
-        # Guillaume's implementation relies on figure*[p] for dedicated page positioning
-
-    def test_dedicated_page_figures_with_scaling(self):
-        """Test Guillaume's specific scaling issue with dedicated page figures.
-
-        Guillaume reported: tex_position="p" works with width=textwidth, but when using
-        other widths like 0.8 or 80%, the figure reverts to 2-column mode.
-        """
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test cases for Guillaume's scaling scenarios
-        test_cases = [
-            # Guillaume's working case - ALL dedicated page figures use figure*[p] to span full width
-            {
-                "width": "\\textwidth",
-                "tex_position": "p",
-                "expected_env": "figure*",
-                "expected_pos": "[p]",
-                "description": "textwidth with position p should use figure*[p] for dedicated page full-width",
-            },
-            # Guillaume's problematic cases that should now work
-            {
-                "width": "0.8",
-                "tex_position": "p",
-                "expected_env": "figure*",
-                "expected_pos": "[p]",
-                "description": "0.8 width with position p should use figure*[p] for dedicated page full-width",
-            },
-            {
-                "width": "80%",
-                "tex_position": "p",
-                "expected_env": "figure*",
-                "expected_pos": "[p]",
-                "description": "80% width with position p should use figure*[p] for dedicated page full-width",
-            },
-            {
-                "width": "0.9\\textwidth",
-                "tex_position": "p",
-                "expected_env": "figure*",
-                "expected_pos": "[p]",
-                "description": "0.9textwidth with position p should use figure*[p] for dedicated page full-width",
-            },
-            # Verify that 2-column still works when no explicit positioning
-            {
-                "width": "\\textwidth",
-                "expected_env": "figure*",
-                "expected_pos": "[!tbp]",
-                "description": "textwidth without explicit positioning should auto-detect 2-column",
-            },
-        ]
-
-        for case in test_cases:
-            attributes = {k: v for k, v in case.items() if k not in ["expected_env", "expected_pos", "description"]}
-
-            result = create_latex_figure_environment(
-                path="FIGURES/test_scaling.svg", caption="Test figure for scaling", attributes=attributes
-            )
-
-            expected_start = f"\\begin{{{case['expected_env']}}}{case['expected_pos']}"
-            assert expected_start in result, (
-                f"Failed for {case['description']}: "
-                f"expected '{expected_start}' in result. "
-                f"Attributes: {attributes}. "
-                f"Got: {result[:200]}..."
-            )
-
-            # For dedicated page figures, verify it's using figure* and not figure
-            if case.get("tex_position") == "p":
-                wrong_start = "\\begin{figure}[p]"
-                assert wrong_start not in result, (
-                    f"Failed for {case['description']}: dedicated page figures should use figure*[p], not figure[p]. Attributes: {attributes}"
-                )
-
-    def test_end_to_end_tex_generation_with_guillaume_fixes(self):
-        """End-to-end test that generates actual .tex file to verify Guillaume's fixes work in practice."""
-        import os
-        import tempfile
-        from pathlib import Path
-
-        with tempfile.TemporaryDirectory() as tmpdir:
-            tmpdir_path = Path(tmpdir)
-            original_cwd = os.getcwd()
-            os.chdir(tmpdir)
-
-            try:
-                # Create manuscript structure with Introduction section
-                manuscript_dir = tmpdir_path / "TEST_MANUSCRIPT"
-                manuscript_dir.mkdir()
-
-                # Create main manuscript file with Introduction section
-                main_md = manuscript_dir / "01_MAIN.md"
-                main_md.write_text("""---
-title:
-  long: "Test Article with Introduction Section"
-  short: "Test Article"
-authors:
-  - name: "Test Author"
-    affiliation: "Test University"
-keywords: ["test", "Guillaume", "fixes"]
-acknowledge_rxiv_maker: false
----
-
-# Abstract
-
-This is the abstract.
-
-## Introduction
-
-This is the introduction content that should appear under "Introduction" header, not "Main".
-
-## Methods
-
-This is the methods section.
-
-## Results
-
-These are the results.
-""")
-
-                # Create YAML front matter file
-                yaml_file = manuscript_dir / "00_FRONT_MATTER.yaml"
-                yaml_file.write_text("""
-title:
-  long: "Test Article with Introduction Section"
-  short: "Test Article"
-authors:
-  - name: "Test Author"
-    affiliation: "Test University"
-keywords: ["test", "Guillaume", "fixes"]
-acknowledge_rxiv_maker: false
-""")
-
-                # Create FIGURES directory with ready file
-                figures_dir = manuscript_dir / "FIGURES"
-                figures_dir.mkdir()
-                ready_fig = figures_dir / "TestFig.png"
-                ready_fig.write_text("fake png content")
-
-                # Create a figure with Guillaume's specific positioning case
-                figures_md = """
-![Test Figure with Ready File](FIGURES/TestFig.png)
-
-![](FIGURES/TestFig.png){#fig:fullpage width="\\textwidth" tex_position="p"}
-**This figure should be on a dedicated page, not 2-column layout.**
-"""
-
-                # Add figures to main content
-                current_content = main_md.read_text()
-                main_md.write_text(current_content + "\n\n" + figures_md)
-
-                # Generate the manuscript using the actual CLI
-                os.environ["MANUSCRIPT_PATH"] = str(manuscript_dir)
-
-                # Change to manuscript directory like real usage
-                os.chdir(manuscript_dir)
-
-                # Import and use the actual generation functions
-                from rxiv_maker.engines.operations.generate_preprint import generate_preprint
-                from rxiv_maker.processors.yaml_processor import extract_yaml_metadata
-
-                # Generate output
-                output_dir = tmpdir_path / "output"
-                output_dir.mkdir()
-
-                # Extract metadata and generate preprint
-                yaml_metadata = extract_yaml_metadata(str(main_md))
-                tex_file = generate_preprint(str(output_dir), yaml_metadata)
-
-                # Verify the generated .tex file contains Guillaume's fixes
-                tex_content = Path(tex_file).read_text()
-
-                # 1. Should have Introduction section, not Main
-                assert "\\section*{Introduction}" in tex_content, (
-                    "Generated .tex should contain Introduction section header"
-                )
-                assert "This is the introduction content" in tex_content, (
-                    "Generated .tex should contain introduction content"
-                )
-                # Should NOT have hardcoded Main section when Introduction exists
-                assert tex_content.count("\\section*{Main}") == 0, (
-                    "Generated .tex should NOT contain Main section when Introduction exists"
-                )
-
-                # 2. Should use ready file path for TestFig
-                assert "Figures/TestFig.png" in tex_content, "Generated .tex should use ready file path directly"
-                assert "Figures/TestFig/TestFig.png" not in tex_content, (
-                    "Generated .tex should NOT use subdirectory format for ready files"
-                )
-
-                # 3. Should use figure* environment for full-page textwidth figures to span full width
-                # Look for the pattern of a figure with tex_position="p" and width=textwidth
-                import re
-
-                fullpage_pattern = (
-                    r"\\begin{figure\*}\[p\].*?width=\\textwidth.*?This figure should be on a dedicated page"
-                )
-                assert re.search(fullpage_pattern, tex_content, re.DOTALL), (
-                    "Generated .tex should use figure*[p] for dedicated page textwidth figures to span full width"
-                )
-                # Should use figure*[p] for dedicated page placement
-                assert "\\begin{figure*}[p]" in tex_content, (
-                    "Generated .tex should use figure*[p] for dedicated page placement"
-                )
-
-                print(f"✅ End-to-end test passed! Generated .tex file: {tex_file}")
-                print("✅ All Guillaume's fixes verified in actual .tex generation")
-
-            finally:
-                os.chdir(original_cwd)
-                if "MANUSCRIPT_PATH" in os.environ:
-                    del os.environ["MANUSCRIPT_PATH"]
-
-
-class TestGuillaumeEdgeCases:
-    """Test edge cases for Guillaume's issues to prevent regressions."""
-
-    def test_mixed_ready_and_generated_figures(self):
-        """Test manuscripts with both ready files and generated figures."""
-        import os
-        import tempfile
-        from pathlib import Path
-
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        with tempfile.TemporaryDirectory() as tmpdir:
-            tmpdir_path = Path(tmpdir)
-            original_cwd = os.getcwd()
-            original_manuscript_path = os.getenv("MANUSCRIPT_PATH")
-
-            try:
-                os.environ["MANUSCRIPT_PATH"] = str(tmpdir_path)
-                os.chdir(tmpdir)
-
-                # Create FIGURES directory
-                figures_dir = tmpdir_path / "FIGURES"
-                figures_dir.mkdir()
-
-                # Create one ready file, leave another as generated
-                ready_file = figures_dir / "ReadyFig.png"
-                ready_file.write_text("ready figure content")
-
-                # Test ready file
-                ready_latex = create_latex_figure_environment(
-                    path="FIGURES/ReadyFig.png", caption="Ready figure", attributes={}
-                )
-                assert "Figures/ReadyFig.png" in ready_latex, "Ready file should use direct path"
-
-                # Test generated file (no ready file exists)
-                generated_latex = create_latex_figure_environment(
-                    path="FIGURES/GeneratedFig.png", caption="Generated figure", attributes={}
-                )
-                assert "Figures/GeneratedFig.png" in generated_latex, (
-                    "Generated file should use direct format (Guillaume's implementation)"
-                )
-
-            finally:
-                os.chdir(original_cwd)
-                if original_manuscript_path is not None:
-                    os.environ["MANUSCRIPT_PATH"] = original_manuscript_path
-                elif "MANUSCRIPT_PATH" in os.environ:
-                    del os.environ["MANUSCRIPT_PATH"]
-
-    def test_panel_references_edge_cases(self):
-        """Test panel reference edge cases."""
-        from rxiv_maker.converters.figure_processor import convert_figure_references_to_latex
-
-        test_cases = [
-            # Multiple panels in sequence - note the {} prevents unwanted spaces after \ref{}
-            (
-                "(@fig:test A), (@fig:test B), (@fig:test C)",
-                "Fig. \\ref{fig:test}{}A), (Fig. \\ref{fig:test}{}B), (Fig. \\ref{fig:test}{}C)",
-            ),
-            # Mixed with other text
-            (
-                "As shown in (@fig:test A) and described in (@fig:test B), the results...",
-                "Fig. \\ref{fig:test}{}A) and described in (Fig. \\ref{fig:test}{}B)",
-            ),
-            # Different figure IDs
-            ("(@fig:first A) vs (@fig:second B)", "Fig. \\ref{fig:first}{}A) vs (Fig. \\ref{fig:second}{}B)"),
-            # Supplementary figures
-            ("(@sfig:sup A) and (@sfig:sup B)", "Fig. \\ref{sfig:sup}{}A) and (Fig. \\ref{sfig:sup}{}B)"),
-        ]
-
-        for input_text, expected_pattern in test_cases:
-            result = convert_figure_references_to_latex(input_text)
-            assert expected_pattern in result, f"Failed for: {input_text} -> {result}"
-
-    def test_complex_positioning_combinations(self):
-        """Test complex combinations of figure positioning and width."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        test_cases = [
-            # Guillaume's specific case - dedicated page figures use figure*[p] for full-width spanning
-            {"width": "\\textwidth", "tex_position": "p", "expected_env": "figure*", "expected_pos": "[p]"},
-            # Two-column spanning variations
-            {"width": "\\textwidth", "expected_env": "figure*", "expected_pos": "[!tbp]"},
-            {"width": "\\textwidth", "tex_position": "t", "expected_env": "figure*", "expected_pos": "[t]"},
-            {"width": "\\textwidth", "tex_position": "b", "expected_env": "figure*", "expected_pos": "[b]"},
-            # Regular figures - dedicated page figures use figure*[p] for full-width spanning
-            {"width": "0.8", "tex_position": "p", "expected_env": "figure*", "expected_pos": "[p]"},
-            {"width": "\\linewidth", "expected_env": "figure", "expected_pos": "[!htbp]"},
-        ]
-
-        for case in test_cases:
-            attributes = {k: v for k, v in case.items() if k not in ["expected_env", "expected_pos"]}
-            result = create_latex_figure_environment(
-                path="FIGURES/test.png", caption="Test caption", attributes=attributes
-            )
-
-            expected_start = f"\\begin{{{case['expected_env']}}}{case['expected_pos']}"
-            assert expected_start in result, f"Failed for {attributes}: expected {expected_start} in {result}"
-
-    def test_manuscript_path_edge_cases(self):
-        """Test edge cases for manuscript path resolution."""
-        import os
-        import tempfile
-        from pathlib import Path
-
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        with tempfile.TemporaryDirectory() as tmpdir:
-            tmpdir_path = Path(tmpdir)
-            original_cwd = os.getcwd()
-            original_manuscript_path = os.getenv("MANUSCRIPT_PATH")
-
-            try:
-                # Test with nested manuscript structure
-                nested_manuscript = tmpdir_path / "project" / "manuscript"
-                nested_manuscript.mkdir(parents=True)
-                figures_dir = nested_manuscript / "FIGURES"
-                figures_dir.mkdir()
-
-                ready_file = figures_dir / "NestedFig.png"
-                ready_file.write_text("nested figure content")
-
-                # Test with MANUSCRIPT_PATH pointing to nested directory
-                os.environ["MANUSCRIPT_PATH"] = str(nested_manuscript)
-                os.chdir(tmpdir)  # Different from manuscript directory
-
-                result = create_latex_figure_environment(
-                    path="FIGURES/NestedFig.png", caption="Nested figure", attributes={}
-                )
-
-                assert "Figures/NestedFig.png" in result, "Should handle nested manuscript paths"
-
-            finally:
-                os.chdir(original_cwd)
-                if original_manuscript_path is not None:
-                    os.environ["MANUSCRIPT_PATH"] = original_manuscript_path
-                elif "MANUSCRIPT_PATH" in os.environ:
-                    del os.environ["MANUSCRIPT_PATH"]
-
-
-class TestGuillaumePR131Rewrite:
-    """Test specific features introduced in Guillaume's PR #131 figure processor rewrite.
-
-    PR #131 completely rewrote the create_latex_figure_environment function with:
-    - Inline figure support
-    - Enhanced width parsing (percentages, fractions, LaTeX units)
-    - Landscape orientation support
-    - Better error handling and attribute parsing
-    - Improved positioning logic for dedicated pages
-    """
-
-    def test_inline_figure_support(self):
-        """Test the new inline=true attribute for non-floating figures."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test inline figure with basic attributes
-        result = create_latex_figure_environment(
-            path="FIGURES/test.png",
-            caption="Test inline figure caption",
-            attributes={"inline": True, "width": "0.8", "id": "fig:inline"},
-        )
-
-        # Should use center environment, not figure
-        assert "\\begin{center}" in result, "Inline figures should use center environment"
-        assert "\\end{center}" in result, "Should properly close center environment"
-        assert "\\begin{figure" not in result, "Inline figures should NOT use figure environment"
-
-        # Should use captionof for local caption
-        assert "\\captionof{figure}" in result, "Inline figures should use captionof for local captions"
-        assert "\\label{fig:inline}" in result, "Should include label when provided"
-
-        # Should have raggedright justification (not center)
-        assert "justification=raggedright" in result, "Inline captions should be left-aligned"
-        assert "singlelinecheck=false" in result, "Should disable single line centering"
-
-    def test_enhanced_width_parsing_percentages(self):
-        """Test enhanced width parsing for percentage values."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        test_cases = [
-            ("80%", "0.800\\linewidth"),
-            ("50%", "0.500\\linewidth"),
-            ("100%", "1.000\\linewidth"),
-            ("75%", "0.750\\linewidth"),
-        ]
-
-        for input_width, expected_width in test_cases:
-            result = create_latex_figure_environment(
-                path="FIGURES/test.png", caption="Test caption", attributes={"width": input_width}
-            )
-
-            assert expected_width in result, (
-                f"Width '{input_width}' should be parsed to '{expected_width}' but got: {result}"
-            )
-
-    def test_enhanced_width_parsing_fractions(self):
-        """Test enhanced width parsing for decimal fractions."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        test_cases = [
-            ("0.8", "0.800\\linewidth"),
-            ("0.5", "0.500\\linewidth"),
-            ("1.0", "1.000\\linewidth"),
-            ("0.75", "0.750\\linewidth"),
-        ]
-
-        for input_width, expected_width in test_cases:
-            result = create_latex_figure_environment(
-                path="FIGURES/test.png", caption="Test caption", attributes={"width": input_width}
-            )
-
-            assert expected_width in result, (
-                f"Width '{input_width}' should be parsed to '{expected_width}' but got: {result}"
-            )
-
-    def test_enhanced_width_parsing_latex_units(self):
-        """Test enhanced width parsing for various LaTeX units."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test LaTeX relative units (these get processed correctly)
-        test_cases = [
-            ("0.8\\textwidth", "0.800\\textwidth"),
-            ("0.9\\columnwidth", "0.900\\columnwidth"),
-        ]
-
-        for input_width, expected_width in test_cases:
-            result = create_latex_figure_environment(
-                path="FIGURES/test.png", caption="Test caption", attributes={"width": input_width}
-            )
-
-            assert expected_width in result, (
-                f"Width '{input_width}' should be parsed to '{expected_width}' but got: {result}"
-            )
-
-        # Test absolute units - these get clamped to \linewidth due to safety clamp for single-column figures
-        absolute_test_cases = ["10cm", "5in", "200pt"]
-
-        for input_width in absolute_test_cases:
-            result = create_latex_figure_environment(
-                path="FIGURES/test.png", caption="Test caption", attributes={"width": input_width}
-            )
-            # Absolute units in single-column figures get safety-clamped to \linewidth
-            assert "width=\\linewidth" in result, (
-                f"Absolute width '{input_width}' should be safety-clamped to \\linewidth for single-column figures"
-            )
-
-        # Test absolute units with strict_width=true - should preserve exact width
-        for input_width in absolute_test_cases:
-            result = create_latex_figure_environment(
-                path="FIGURES/test.png", caption="Test caption", attributes={"width": input_width, "strict_width": True}
-            )
-            assert f"width={input_width}" in result, (
-                f"Absolute width '{input_width}' with strict_width=true should preserve exact width"
-            )
-
-    def test_landscape_figure_support(self):
-        """Test the new landscape=true attribute for sideways figures."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test regular landscape figure
-        result = create_latex_figure_environment(
-            path="FIGURES/wide_plot.png",
-            caption="Wide landscape plot",
-            attributes={"landscape": True, "id": "fig:landscape"},
-        )
-
-        assert "\\begin{sidewaysfigure}" in result, "Landscape figures should use sidewaysfigure"
-        assert "\\end{sidewaysfigure}" in result, "Should properly close sidewaysfigure"
-        assert "\\begin{figure}" not in result, "Should not use regular figure environment"
-
-        # Test landscape with two-column spanning
-        result2 = create_latex_figure_environment(
-            path="FIGURES/wide_plot.png",
-            caption="Wide landscape plot spanning columns",
-            attributes={"landscape": True, "width": "\\textwidth", "id": "fig:landscape2"},
-        )
-
-        assert "\\begin{sidewaysfigure*}" in result2, "Two-column landscape should use sidewaysfigure*"
-        assert "\\end{sidewaysfigure*}" in result2, "Should properly close sidewaysfigure*"
-
-    def test_barrier_support(self):
-        """Test the new barrier=true attribute for float barriers."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        result = create_latex_figure_environment(
-            path="FIGURES/test.png", caption="Figure with barrier", attributes={"barrier": True}
-        )
-
-        assert "\\FloatBarrier" in result, "Barrier=true should add FloatBarrier command"
-
-        # Test without barrier
-        result2 = create_latex_figure_environment(
-            path="FIGURES/test.png", caption="Figure without barrier", attributes={}
-        )
-
-        assert "\\FloatBarrier" not in result2, "Should not add FloatBarrier by default"
-
-    def test_max_height_attribute(self):
-        """Test the new max_height attribute for height constraints."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        result = create_latex_figure_environment(
-            path="FIGURES/tall_figure.png",
-            caption="Tall figure with height constraint",
-            attributes={"width": "\\textwidth", "max_height": "0.8\\textheight"},
-        )
-
-        assert "height=0.800\\textheight" in result, "Should include height constraint"
-        assert "width=\\textwidth" in result, "Should still include width"
-        assert "keepaspectratio" in result, "Should maintain aspect ratio with both width and height"
-
-    def test_fit_presets(self):
-        """Test the new fit attribute with page/width/height presets."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test fit=page with dedicated page positioning
-        result_page = create_latex_figure_environment(
-            path="FIGURES/full_page.png", caption="Full page figure", attributes={"fit": "page", "tex_position": "p"}
-        )
-
-        assert "width=\\textwidth" in result_page, "fit=page should use full textwidth"
-        assert "height=0.95\\textheight" in result_page, "fit=page should use most of textheight"
-        assert "\\begin{figure*}[p]" in result_page, "fit=page with position=p should use figure*[p]"
-
-        # Test fit=width
-        result_width = create_latex_figure_environment(
-            path="FIGURES/wide.png", caption="Full width figure", attributes={"fit": "width"}
-        )
-
-        assert "width=\\linewidth" in result_width, "fit=width should use full linewidth"
-
-        # Test fit=height
-        result_height = create_latex_figure_environment(
-            path="FIGURES/tall.png", caption="Full height figure", attributes={"fit": "height"}
-        )
-
-        assert "height=\\textheight" in result_height, "fit=height should use full textheight"
-
-    def test_fullpage_attribute_alias(self):
-        """Test that fullpage=true is equivalent to fit=page."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        result = create_latex_figure_environment(
-            path="FIGURES/fullpage.png",
-            caption="Full page figure using fullpage attribute",
-            attributes={"fullpage": True, "tex_position": "p"},
-        )
-
-        assert "width=\\textwidth" in result, "fullpage=true should use full textwidth"
-        assert "height=0.95\\textheight" in result, "fullpage=true should use most of textheight"
-        assert "\\begin{figure*}[p]" in result, "fullpage=true should use figure*[p] for dedicated page"
-
-    def test_improved_position_parsing(self):
-        """Test improved positioning attribute parsing."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test positions with brackets are properly stripped
-        test_cases = [
-            ("[!htbp]", "[!htbp]"),
-            ("!htbp", "[!htbp]"),
-            ("[tp]", "[tp]"),
-            ("tp", "[tp]"),
-            ("[p]", "[p]"),
-            ("p", "[p]"),
-        ]
-
-        for input_pos, expected_pos in test_cases:
-            result = create_latex_figure_environment(
-                path="FIGURES/test.png", caption="Test positioning", attributes={"tex_position": input_pos}
-            )
-
-            # For dedicated page positioning, should use figure*
-            if "[p]" in expected_pos:
-                assert f"\\begin{{figure*}}{expected_pos}" in result, (
-                    f"Position '{input_pos}' should parse to '{expected_pos}' and use figure* for dedicated page"
-                )
-            else:
-                assert f"\\begin{{figure}}{expected_pos}" in result, (
-                    f"Position '{input_pos}' should parse to '{expected_pos}' in regular figure"
-                )
-
-    def test_caption_width_customization(self):
-        """Test the new caption_width attribute for custom caption widths."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        result = create_latex_figure_environment(
-            path="FIGURES/test.png",
-            caption="Figure with custom caption width",
-            attributes={"width": "\\textwidth", "caption_width": "0.8\\textwidth"},
-        )
-
-        assert "width=0.8\\textwidth" in result, "Should use custom caption width"
-        assert "\\captionsetup{" in result, "Should have caption setup"
-
-    def test_strict_width_attribute(self):
-        """Test the new strict_width attribute to prevent width clamping."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test without strict_width - should clamp textwidth to linewidth for single-column
-        result1 = create_latex_figure_environment(
-            path="FIGURES/test.png",
-            caption="Regular figure",
-            attributes={"width": "\\textwidth"},  # No strict_width, should get clamped to linewidth
-        )
-
-        # With textwidth, should auto-upgrade to figure* for proper two-column spanning
-        assert "\\begin{figure*}" in result1, "textwidth should auto-upgrade to figure* for spanning"
-
-        # Test with strict_width=true - should preserve exact width
-        result2 = create_latex_figure_environment(
-            path="FIGURES/test.png",
-            caption="Strict width figure",
-            attributes={"width": "\\textwidth", "strict_width": True},
-        )
-
-        assert "width=\\textwidth" in result2, "strict_width=true should preserve exact width"
-        assert "\\begin{figure*}" in result2, "textwidth should still use figure* for proper rendering"
-
-    def test_error_handling_graceful_fallbacks(self):
-        """Test that PR #131's error handling provides graceful fallbacks."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test with invalid width value - should still work
-        result = create_latex_figure_environment(
-            path="FIGURES/test.png", caption="Figure with invalid width", attributes={"width": "invalid_width_value"}
-        )
-
-        # Should still generate valid LaTeX even with invalid width
-        assert "\\begin{figure}" in result, "Should generate valid figure even with invalid width"
-        assert "\\includegraphics" in result, "Should include graphics command"
-        assert "\\caption{Figure with invalid width}" in result, "Should include caption"
-
-    def test_complex_attribute_combinations(self):
-        """Test complex combinations of new PR #131 attributes."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test landscape + barrier + custom positioning
-        result1 = create_latex_figure_environment(
-            path="FIGURES/complex1.png",
-            caption="Complex landscape figure",
-            attributes={
-                "landscape": True,
-                "barrier": True,
-                "tex_position": "!t",
-                "width": "0.9\\textwidth",
-                "id": "fig:complex1",
-            },
-        )
-
-        assert "\\begin{sidewaysfigure*}[!t]" in result1, "Should combine landscape + two-column + positioning"
-        assert "\\FloatBarrier" in result1, "Should include barrier"
-        assert "width=0.900\\textwidth" in result1, "Should parse width correctly"
-
-        # Test inline + max_height + custom caption width
-        result2 = create_latex_figure_environment(
-            path="FIGURES/complex2.png",
-            caption="Complex inline figure with height constraint",
-            attributes={"inline": True, "width": "80%", "max_height": "10cm", "id": "fig:complex2"},
-        )
-
-        assert "\\begin{center}" in result2, "Should use center for inline"
-        assert "width=0.800\\linewidth" in result2, "Should parse percentage width"
-        assert "height=10cm" in result2, "Should include height constraint"
-        assert "\\captionof{figure}" in result2, "Should use captionof for inline"
-
-    def test_pr131_regression_prevention(self):
-        """Test that PR #131 changes don't break existing functionality."""
-        from rxiv_maker.converters.figure_processor import create_latex_figure_environment
-
-        # Test basic figure (should work exactly as before)
-        basic_result = create_latex_figure_environment(path="FIGURES/basic.png", caption="Basic figure", attributes={})
-
-        assert "\\begin{figure}[!htbp]" in basic_result, "Basic figures should still work"
-        assert "width=\\linewidth" in basic_result, "Should default to linewidth"
-        assert "\\includegraphics" in basic_result, "Should include graphics"
-        assert "\\caption{Basic figure}" in basic_result, "Should include caption"
-
-        # Test textwidth auto-detection (key Guillaume fix)
-        textwidth_result = create_latex_figure_environment(
-            path="FIGURES/wide.png", caption="Wide figure", attributes={"width": "\\textwidth"}
-        )
-
-        assert "\\begin{figure*}" in textwidth_result, "textwidth should auto-upgrade to figure*"
-        assert "width=\\textwidth" in textwidth_result, "Should preserve textwidth"
-
-        # Test dedicated page positioning (key Guillaume fix)
-        dedicated_result = create_latex_figure_environment(
-            path="FIGURES/dedicated.png",
-            caption="Dedicated page figure",
-            attributes={"tex_position": "p", "width": "\\textwidth"},
-        )
-
-        assert "\\begin{figure*}[p]" in dedicated_result, "Dedicated page should use figure*[p]"
-        assert "width=\\textwidth" in dedicated_result, "Should use full textwidth for dedicated page"
-
-
-if __name__ == "__main__":
-    pytest.main([__file__])
